+++ frontend/src/component/admin/role/CreateRoleComponent.tsx
@@ -1,23 +1,24 @@
 import {useState} from 'react';
-import {Button, Flex, Form, Input} from 'antd';
+import {Button, Collapse, Flex, Form, Input} from 'antd';
 import {AdminService} from "../../../api/AdminService.tsx";
 import {useNavigate} from "react-router-dom";
 import {Routes} from "../../../models/Routes.tsx";
 import {MdSave} from "react-icons/md";
 import CardHeader from "../../common/CardHeader.tsx";
 import {NotificationHandler} from "../../../util/NotificationHandler.tsx";
+import ScopeComponent from "./scopes/ScopeComponent.tsx";
 
 const CreateRoleCOmponent = () => {
     const navigate = useNavigate();
     const [form] = Form.useForm();
-
     const [loading, isLoading] = useState(false);
+    const [scopes, setScopes] = useState<string[]>([]);
 
     //@ts-ignore
     const handleSubmit = async (values) => {
         try {
             isLoading(true);
-            await AdminService.Role.createRole(values.name, values.description);
+            await AdminService.Role.createRole(values.name, values.description, scopes);
             navigate(Routes.Admin.RoleSection.LIST);
         } catch (e) {
             NotificationHandler.showErrorNotificationFromError(e as Error);
@@ -47,8 +48,16 @@ const CreateRoleCOmponent = () => {
                         rows={4}
                     />
                 </Form.Item>
+                <Collapse items={[
+                    {
+                        key: '1',
+                        label: 'Scopes',
+                        children: <ScopeComponent scopes={scopes} onScopesSelected={setScopes}/>
+                    }
+                ]}>
+                </Collapse>
                 <Form.Item>
-                    <Flex justify={"flex-end"} align={"flex-end"}>
+                    <Flex justify={"flex-end"} align={"flex-end"} style={{marginTop: '20px'}}>
                         <Button type="primary" htmlType="submit" loading={loading} icon={<MdSave/>}>
                             Add
                         </Button>